+++ driver_mobile/lib/widgets/custom_button.dart
@@ -33,6 +33,14 @@ class CustomButton extends StatelessWidget {
             isLight: isElevated,
             color: isElevated ? AppColors.background : AppColors.primary,
           )
+        : Row(
+            mainAxisSize: MainAxisSize.min,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              if (icon != null) ...[icon!, const SizedBox(width: 8.0)],
+              Text(text),
+            ],
+          );
         : (icon == null
             ? Text(
                 text,
@@ -63,7 +71,9 @@ class CustomButton extends StatelessWidget {
         button = ElevatedButton(
           onPressed: activeOnPressed,
           style: height != null
-              ? ElevatedButton.styleFrom(minimumSize: Size(width ?? double.infinity, height!))
+              ? ElevatedButton.styleFrom(
+                  minimumSize: Size(width ?? double.infinity, height!),
+                )
               : null,
           child: label,
         );
@@ -72,7 +82,9 @@ class CustomButton extends StatelessWidget {
         button = OutlinedButton(
           onPressed: activeOnPressed,
           style: height != null
-              ? OutlinedButton.styleFrom(minimumSize: Size(width ?? double.infinity, height!))
+              ? OutlinedButton.styleFrom(
+                  minimumSize: Size(width ?? double.infinity, height!),
+                )
               : null,
           child: label,
         );
@@ -81,7 +93,9 @@ class CustomButton extends StatelessWidget {
         button = TextButton(
           onPressed: activeOnPressed,
           style: height != null
-              ? TextButton.styleFrom(minimumSize: Size(width ?? double.infinity, height!))
+              ? TextButton.styleFrom(
+                  minimumSize: Size(width ?? double.infinity, height!),
+                )
               : null,
           child: label,
         );
@@ -89,10 +103,7 @@ class CustomButton extends StatelessWidget {
     }
 
     if (width != null) {
-      return SizedBox(
-        width: width,
-        child: button,
-      );
+      return SizedBox(width: width, child: button);
     }
     return button;
   }